django-learn/django_vue1/backend/requirements.txt
django_filter==2.2.0
six==1.12.0
Django==2.2.10
django_filters==0.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
coreapi==2.3.3
django-cors-headers==3.3.0